+++ tesseraql-camel-runtime/src/test/java/io/tesseraql/runtime/MySqlPortabilityIntegrationTest.java
@@ -91,6 +91,11 @@ class MySqlPortabilityIntegrationTest {
                 "varchar(%d)");
     }
 
+    @Test
+    void theCatalogVersionTableAppliesOnThisDialect() throws Exception {
+        DialectRuntimeChecks.catalogVersionTableApplies(mysqlDataSource());
+    }
+
     private static javax.sql.DataSource mysqlDataSource() {
         com.mysql.cj.jdbc.MysqlDataSource dataSource = new com.mysql.cj.jdbc.MysqlDataSource();
         dataSource.setUrl(MYSQL.getJdbcUrl());